Working liquid tab minimize behaviour
Enable native tab bar minimize-on-scroll on iOS via NativeTabs minimizeBehavior, add a dedicated Test tab with a ScrollView-only screen, and remove the older minimize test route and bottom accessory so the fifth tab registers reliably with expo-router.

### frontend/dailyflo/app/(tabs)/_layout.tsx
- Set minimizeBehavior to onScrollDown on iOS for the whole NativeTabs navigator (documented API).
- Add a fifth trigger named test with label Test and a PNG tab icon matching other tabs.
- Drop MinimizeTestBottomAccessory wiring and segment-based minimize toggling; remove MaterialCommunityIcons import.

### frontend/dailyflo/app/(tabs)/test/index.tsx
- New tab screen: vertical ScrollView with long placeholder content and iOS contentInsetAdjustmentBehavior automatic for safe-area / tab bar coordination.

### frontend/dailyflo/app/(tabs)/minimizeTest/index.tsx
- Removed; replaced by the test route to match Trigger name and file-based routing.

### frontend/dailyflo/app/(tabs)/minimizeTest/_layout.tsx
- Removed; unnecessary wrapper for the old minimizeTest route.

### frontend/dailyflo/app/(tabs)/MinimizeTestBottomAccessory.tsx
- Removed; simplify native-tabs setup (minimize is driven by NativeTabs props and ScrollView, not BottomAccessory).

### frontend/dailyflo/app/(tabs)/today/index.tsx
- Comment only: clarify ordering of list vs backdrop for scroll/layout (no functional change tied to minimize).

Test tab not appearing in the native tab bar:
- ROOT CAUSE: expo-router only links NativeTabs.Trigger names to existing file routes; vector-only tab icons or a mismatched segment (e.g. minimizeTest) can drop the tab or fail to register.
- SOLUTION: Use app/(tabs)/test/index.tsx with Trigger name="test", and the same PNG + renderingMode="template" pattern as other tabs.

Conditional minimizeBehavior only on a single tab:
- ROOT CAUSE: NativeTabs minimizeBehavior is a navigator-level prop; per-tab toggling is easy to get wrong and was removed in favor of one iOS-wide setting plus ScrollView on the test screen.
- SOLUTION: Apply minimizeBehavior onScrollDown for all iOS tabs; rely on Browse/Test ScrollView surfaces for the behavior the system can track.
